// src/app/Model/student.ts
export class Student {
    user_id: number;
    image: File;
    firstName: string;
    middleName: string;
    lastName: string;
    gender : string;
    mothersName: string;
    mobileNumber: string;
    parentNumber: string;
    address: string;
    city :string;
    state:string;
    pincode:string;
    adharNumber: string;
    dateOfBirth: string;
    higherQualification: string;
    email: string;
    username: string;
    password: string;
    bloodGroup: string;
    role: string;
    type:string;
    courseName:string;
    batchName:string;
    course_id:string;
    batch_id:string;
    
    constructor() {}
}